Move stat calculations to module
