+++ lib/require_helper.rb
@@ -7,5 +7,6 @@ require_relative './team'
 require_relative './team_season'
 require_relative './game'
 require_relative './coach'
+require_relative './stat_calculatable'
 require 'csv'
 require 'benchmark'
